feat: 예매 조회 dto, repo 추가
