+++ src/main/java/com/fifo/ticketing/TicketingApplication.java
@@ -2,10 +2,11 @@ package com.fifo.ticketing;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @EnableJpaAuditing
-@SpringBootApplication
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
 //@EnableScheduling 스케줄링 추가시
 public class TicketingApplication {
