+++ src/main/java/com/fifo/ticketing/domain/book/repository/BookRepository.java
@@ -1,6 +1,7 @@
 package com.fifo.ticketing.domain.book.repository;
 
 import com.fifo.ticketing.domain.book.entity.Book;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookRepository extends JpaRepository<Book, Long> {
