+++ src/main/java/com/fifo/ticketing/global/config/SecurityConfig.java
@@ -14,6 +14,7 @@ import org.springframework.security.web.SecurityFilterChain;
 
 @Configuration
 @RequiredArgsConstructor
+@EnableWebSecurity
 public class SecurityConfig {
 
     private final FormLoginSuccessHandler formLoginSuccessHandler;
